+++ src/it/unibs/fp/rovineperdute/MatrixTeamMetztli.java
@@ -8,7 +8,7 @@ public class MatrixTeamMetztli extends AdjacencyMatrix {
 
 
     @Override
-    public double calculateWeight(City cityA, City cityB) {
+    public float calculateWeight(City cityA, City cityB) {
         return Math.abs(cityA.getCoordinates().getH() - cityB.getCoordinates().getH());
     }
 }
