+++ src/it/unibs/fp/rovineperdute/MatrixTeamTonatiuh.java
@@ -7,11 +7,11 @@ public class MatrixTeamTonatiuh extends AdjacencyMatrix {
     }
 
     @Override
-    public double calculateWeight(City cityA, City cityB) {
+    public float calculateWeight(City cityA, City cityB) {
         double weightX = Math.pow(cityA.getCoordinates().getX() - cityB.getCoordinates().getX(), 2);
         double weightY = Math.pow(cityA.getCoordinates().getY() - cityB.getCoordinates().getY(), 2);
 
-        return Math.sqrt(weightX + weightY);
+        return (float)Math.sqrt(weightX + weightY);
     }
 
 }
